pricewatch_v4/requirements.txt
fastapi
uvicorn[standard]
gunicorn
supabase
pydantic-settings
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
httpx
groq
requests
pytest
pytest-asyncio
passlib[bcrypt]